Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -20,7 +20,7 @@ int main()
     }
     else if (ch == '/')
     {
-        printf("%d", (float) a / b);
+        printf("%d", (int) a / b);
     }
     return 0;
 }
